Added a toolbar to the detail activity.
Added the views needed to display the poster details.

Enabled support for data binding.

Added the release date and user rating labels.

Added a content description for the movie poster image view.

Added constraints to the toolbar

Updated the layout to support scrolling.

Adjusted the title, release date and user rating text view to prevent
running off the of screen.