Added a toolbar to the main layout to serve as an action bar.
Added a gridview to the layout for displaying the movie posters.

Ensure that the height on the views is set to either 0dp or wrap_content
for constraint layout children.

Constrain the gridview on the left, right, and bottom.

Enabled support for data binding.

Cleaned up the layout for the gridview so that it matches the layout of
the mockup.

Added constraints to the toolbar.

Added a network error text view for extreme cases where there is no
network connectivity.

Updated the grid view to user the grid count value from the integers
resource.