Added a toolbar to the main activity which will serve as an action bar.
Overrode onCreateOptionsMenu to create the menu.

Overrode onOptionsItemSelected to show a popup menu for the sort by
requirements.

Thanks to the following link for tips on showing a popup menu on the
action bar.

https://stackoverflow.com/questions/14729592/show-popup-menu-on-actionbar-item-click

Added the MovieListResultAsyncTask AsyncTask to get the json from the
movie database.

Updated MainActivity to use the API key from the PopularMoviesConstants
class.

Updated MainActivity to use the PopularMoviesSettings.

Implemented the sort by menu items.

Handle movie poster clicks and launch the detail activity.

Removed calls to findViewById and use data binding instead.

Save the sort by value in save instance state.

Use the string resource keys for the intent put extra.

Use the string resource key for the activity bundle data.

Removed an erroneous check of null on the results list since the list
will never be null. The check is for empty instead.

Fixed duplicate data being displayed in the grid which was caused by
another async task executing the same page.

Refactored onMenuItemClick to remove duplicate code.

Saved the position of the grid view prior to a configuration change so
that the position can be restored after the configuration changes.

Use the movie_visible_position_key from strings.xml