Added a new string resource for the sort icon.
Added sort by "most popular" and "top rated" strings for the sort by
menu.

Added the release date and user rating labels.

Added a content description string for the movie poster image view.

Added keys for intent put extra

Added a key for the activity bundle data.

Added string resources for the movie database API.

Added a string for the title of the detail activity.

Added a new string movie_visible_position_key to store the position of
the grid view prior to a configuration change.