Added the internet permission.
Added the network state permission so that the app can receive updates
from the system on network connectivity.

Changed the launch mode of the main activity to single top so that when
the user clicks the back button the parent activity is not recreated the
view is in the same position prior to launching a child activity.