Added helper functions for the two new MovieReviewResultObject and MovieVideoResultObject objects.
Fixed a bug when processing the JSON in the
movieReviewJsonStringToMovieReviewResultList and
movieVideoJsonStringToMovieVideoResultList functions. The bug resulted
in processing the *ObjectList instead of the resultsArray.

Fixed a bug when processing the JSON in the
movieReviewJsonStringToMovieReviewResultList and
movieVideoJsonStringToMovieVideoResultList functions. The id field in
the JSON is not the movie identifier.

Both functions have been updated to take the actual movie id as a
parameter.