Added the id field to the intent for the detail activity.
Removed the MovieListResultAsyncTask inner class and use the external
class.

Added an inner AsyncTaskCompletion listener to handle the post execution
of the async task.

Updated the dispatchMovieListResultRequest() to handle movie favorite
database fetches.

Updated the hideGridViewAndShowNetworkErrorMessage and
showGridViewAndHideNetworkErrorMessage functions to handle being in the
favorites state.

Added a new helper function launchBroadCastReceiverIfNotRegistered.

Ensure the broadcast receiver is launched if we change to the "popular"
or "top movie" states.

Display an error message if the favorites database is empty.

Updated the onMenuItemClickHelper function to restore the view state so
that error messages are cleared and the gridview is displayed.

Implement onResume to handle coming back to the main activity from the
detail activity when a favorite movie is added or removed.

Use the shared preferences favorites key from the string resource.