Added a separator between trailers and synopsis.
Added a Trailers text view.

Added a recyclerview to the layout

Prevent the recycler view from stealing focus.

Thanks: https://stackoverflow.com/questions/36923948/recycleview-steals-focus-when-inside-a-nestedscrollview

Added a separator between the trailer recycler view and the reviews text view

Added a reviews text view.

Added an informational text view that is displayed only when there are
no trailers available.

Added an informational text view that is displayed only when there are
no reviews available.

Fixed issues where top_toBottom and bottom_toTop constraints were
applied between each view. This caused large gaps between views. When
views are in a vertical chain, only bottom_toTop constraints are needed
except at the ends of the chain.

Clean up margins and constraints.

Added a "Favorite" button to layout.

Fixed the layout so that the toolbar doesn't scroll along with the rest
of the content.