Initialized the movie trailer recycler view.
Initialized the movie review recycler view.

Added a MovieTrailerResultAsyncTaskCompleteListener inner class to
handle post execution of the trailer async task.

Launch the trailer async task.

Passed a movie review list to the movie review adapter.

Created a completion listener for the movie review async task.

Launched the movie review async task.

Implemented the movie trailer click listener.

Added an inner class to handle the favorite button click.

Added an implicit intent to launch trailers on youtube when a trailer
list item is clicked.

Added an implicit intent to launch trailers on the youtube app when a
trailer list item is clicked.

Created the MovieFavoriteDbQueryAsyncTaskCompleteListener for handling
movie favorite database queries.

Ensure the "favorite" button is not initially visible until the query
task completes.

Added flags to check whether the tasks have been completed.

Added a flag to determine the current state of the favorite button.

Implemented the add and delete portion of the favorite button.

Fixed a bug in the async task query handler that was caused by not
moving the cursor to the first row.

Close the cursor in the query handler when the cursor is no longer
needed.

Notify the main activity of a change in favorites via shared
preferences.

Added a notification flag that is set when the movie poster is
downloaded and loaded into the image view.

Added a helper function to convert the movie poster image view to a byte
array.
Thanks - https://stackoverflow.com/questions/37779515/how-can-i-convert-an-imageview-to-byte-array-in-android-studio

Added a helper function to either use the raw movie poster image data
from the database or fetch the image from the remote server.
Thanks - http://www.tutorialforandroid.com/2009/10/how-to-insert-image-data-to-sqlite.html

Override the create menu functions to add the share menu button. Call
the onShareButtonClick helper function when the share button is clicked.

Show the share button if and only if we have at least one trailer.
Thanks - https://stackoverflow.com/questions/10692755/how-do-i-hide-a-menu-item-in-the-actionbar

Changed the unfavorite field to undo favorite.

Use the youtube app string from strings.xml

Use the youtube website base string from strings.xml