Created the MovieFavoriteContentProvider class.
Added a URI matcher.

Implemented the insert function.

Implemented the query function.

Added a matcher for a single row of data by id.

Added a case in the query function to handle selection by id.

Implemented the delete function.

The update and getType functions are not implemented since we are not
providing this functionality.

Added matchers for the movie review and trailer URIs.

Implemented insert for review and trailer URIs.

Implemented query for review and trailer URIs.

Implemented delete for review and trailer URIs.

Fixed a bug in the query and delete functions where "_id=?" did not
include an equals sign.