Created the MovieFavoriteDbQueryAsyncTask class that performs movie favorite database queries on a background thread.
Fixed a bug in the background function that was caused by not building
up the content uri for the single item case.

Fixed a bug in the doInBackground function caused by checking the params
array for null instead of checking the array length.

Updated to support a generic extra data parameter.

The movieId field is optional so it was changed to an object which is
easier differentiate, i.e. null vs. an invalid integer value.