Only check the favorites shared preferences if we are in the favorites state.
Fixed a bug where the sharedPreferences variable was not initialized.

Fix handling of the UI state on a configuration change. Strangely my
original code did not work until I added a delay. I'm not sure why the
delay is required. Made sure the fix was added for the favorites state
too.